svg/types/elements/SVGGeometryElement-getPointAtLength-exception.html

<!DOCTYPE html>
<title>SVGGeometryElement.getPointAtLength method exception</title>
<link rel="help" href="https://svgwg.org/svg2-draft/types.html#__svg__SVGGeometryElement__getPointAtLength"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  var pathElement = document.createElementNS("http://www.w3.org/2000/svg", "path");
  function pointAtLength(string) {
      pathElement.setAttribute("d", string);
      var point = pathElement.getPointAtLength(700);
      return [Math.round(point.x), Math.round(point.y)];
  }

  assert_throws_dom("InvalidStateError", function() { pointAtLength('M0,20 L400,20 L640,20') });
  assert_throws_dom("InvalidStateError", function() { pointAtLength('M0,20 L400,20 L640,20 z') });
  assert_throws_dom("InvalidStateError", function() { pointAtLength('M0,20 L400,20 z M 320,20 L640,20') });
  assert_throws_dom("InvalidStateError", function() { pointAtLength('M0,20 L20,40') });
}, document.title + " with SVGPathElement");
test(function() {
  var rectElement = document.createElementNS("http://www.w3.org/2000/svg", "rect");
  function pointAtLength(rx, ry, width, height) {
    rectElement.setAttribute("rx", rx);
    rectElement.setAttribute("ry", ry);
    rectElement.setAttribute("width", width);
    rectElement.setAttribute("height", height);
    var point = rectElement.getPointAtLength(300);
    return [Math.round(point.x), Math.round(point.y)];
  }
  assert_throws_dom("InvalidStateError", function() { pointAtLength(0, 0, 200, 300) });
  assert_throws_dom("InvalidStateError", function() { pointAtLength(50, 50, 200, 300) });
}, document.title + " with SVGRectElement");
test(function() {
  var circleElement = document.createElementNS("http://www.w3.org/2000/svg", "circle");
  function pointAtLength(radius) {
    circleElement.setAttribute("r", radius);
    var point = circleElement.getPointAtLength(100);
    return [Math.round(point.x), Math.round(point.y)];
  }
  assert_throws_dom("InvalidStateError", function() { pointAtLength(10) });
  assert_throws_dom("InvalidStateError", function() { pointAtLength(100) });
}, document.title + " with SVGCircleElement");
</script>
